fullstack/python/requirements.txt
attrs==19.3.0
Click==7.0
coverage==5.0.3
factory-boy==2.12.0
Faker==4.0.1
fastjsonschema==2.14.3
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
python-dateutil==2.8.1
six==1.14.0
text-unidecode==1.3
wcwidth==0.1.8
